+++ entities/items/index.js
@@ -1,0 +1,1 @@
+export { ItemManager } from './items.js';